requirements.txt
py-cord
pandas
sklearn
sentence_transformers
python-dotenv